telnyx-twilio-migration/migration-examples/django/appointment-reminders-django/requirements.txt
appdirs==1.4.3
appnope==0.1.0
arrow==0.15.7
asgiref==3.3.2
attrs==19.3.0
backcall==0.1.0
black==19.10b0
certifi==2023.7.22
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
decorator==4.4.2
distlib==0.3.0
Django==3.2.14
django-dramatiq==0.9.1
django-bootstrap3==14.0.0
django-environ==0.4.5
django-forms-bootstrap==3.1.0
django-timezone-field==4.0
dramatiq[rabbitmq]==1.9.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
identify==1.4.11
idna==2.9
importlib-metadata==1.5.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
mccabe==0.6.1
mock==4.0.2
model-mommy==2.0.0
nodeenv==1.3.5
parso==0.6.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.2.0
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pytz==2019.3
PyYAML==6.0.1
redis==3.5.3
regex==2020.2.20
requests==2.23.0
selenium==3.141.0
six==1.16.0
sqlparse==0.3.1
toml==0.10.0
traitlets==4.3.3
telnyx==2.1.6
typed-ast==1.5.4
urllib3==1.25.8
virtualenv==20.0.10
wcwidth==0.1.8
whitenoise==5.1.0
zipp==3.1.0
